+++ microservicios-compras/src/main/java/com/prpymes/microservicios/app/compras/controllers/DetallePedidoController.java
@@ -35,7 +35,7 @@ public class DetallePedidoController extends CommonControllerC<DetallePedido, De
 
 	@Override
 	@GetMapping("/detalle-pedidos/{id}")
-	public ResponseEntity<?> ver(@PathVariable Integer id) {
+	public ResponseEntity<?> ver(@PathVariable Long id) {
 		Optional<DetallePedido> o = service.findById(id);
 		if(!o.isPresent()) {
 			return ResponseEntity.notFound().build();
@@ -59,7 +59,7 @@ public class DetallePedidoController extends CommonControllerC<DetallePedido, De
 
 	@Override
 	@DeleteMapping("/detalle-pedidos/{id}")
-	public ResponseEntity<?> eliminar(@PathVariable Integer id) {
+	public ResponseEntity<?> eliminar(@PathVariable Long id) {
 		service.deleteById(id);
 		return ResponseEntity.noContent().build();
 	}
